+++ src/pages/organisationUnits/form/OrganisationUnitSelector.tsx
@@ -1,7 +1,7 @@
 import i18n from '@dhis2/d2-i18n'
 import { Field, NoticeBox, OrganisationUnitTree } from '@dhis2/ui'
 import { IconInfo16 } from '@dhis2/ui-icons'
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import { useField } from 'react-final-form'
 import { useCurrentUserRootOrgUnits } from '../../../lib/user/currentUserStore'
 import classes from './OrganisationUnitSelector.module.css'
@@ -31,6 +31,7 @@ export function OrganisationUnitSelector() {
 
     return (
         <Field
+            name="parent"
             error={meta.touched && meta.error}
             validationText={meta.touched ? meta.error : undefined}
         >