+++ src/pages/organisationUnits/form/organisationUnitSchema.ts
@@ -1,5 +1,6 @@
+import i18n from '@dhis2/d2-i18n'
 import { z } from 'zod'
-import { getDefaults, modelFormSchemas } from '../../../lib'
+import { createFormValidate, getDefaults, modelFormSchemas } from '../../../lib'
 
 const { withAttributeValues, identifiable, referenceCollection } =
     modelFormSchemas
@@ -9,20 +10,71 @@ export const organisationUnitSchema = identifiable
     .extend({
         shortName: z.string().trim().default(''),
         code: z.string().trim().optional(),
-        description: z.string().trim().optional(),
+        description: z
+            .string()
+            .trim()
+            .max(2147483647, {
+                message: i18n.t('Should not exceed {{maxLength}} characters', {
+                    maxLength: 2147483647,
+                }),
+            })
+            .optional(),
         image: z.object({ id: z.string() }).optional(),
-        phoneNumber: z.string().optional(),
-        contactPerson: z.string().optional(),
-        openingDate: z.string(),
-        email: z.string().optional(),
-        address: z.string().optional(),
-        url: z.string().optional(),
-        closedDate: z.string().optional(),
-        parent: z.object({ id: z.string() }).optional(),
+        phoneNumber: z
+            .string()
+            .min(0, { message: i18n.t('Must be a valid mobile number') })
+            .max(150, { message: i18n.t('Must be a valid mobile number') })
+            .optional(),
+        contactPerson: z
+            .string()
+            .max(255, {
+                message: i18n.t('Should not exceed {{maxLength}} characters', {
+                    maxLength: 255,
+                }),
+            })
+            .optional(),
+        openingDate: z.string().date(),
+        email: z.string().email().optional(),
+        address: z
+            .string()
+            .max(230, {
+                message: i18n.t('Should not exceed {{maxLength}} characters', {
+                    maxLength: 255,
+                }),
+            })
+            .optional(),
+        url: z
+            .string()
+            .url({ message: i18n.t('Must be a valid url') })
+            .optional(),
+        closedDate: z.string().date().optional(),
+        comment: z
+            .string()
+            .max(2147483647, {
+                message: i18n.t('Should not exceed {{maxLength}} characters', {
+                    maxLength: 2147483647,
+                }),
+            })
+            .optional(),
+        parent: z.object({ id: z.string(), path: z.string() }).optional(),
         geometry: z
             .object({
                 type: z.literal('Point'),
-                coordinates: z.array(z.number()).length(2),
+                coordinates: z
+                    .array(z.number())
+                    .length(2)
+                    .refine(
+                        (coord) =>
+                            coord[0] >= -90 &&
+                            coord[0] <= 90 &&
+                            coord[1] >= -180 &&
+                            coord[1] <= 180,
+                        {
+                            message: i18n.t(
+                                'Longitude should be between -90 and 90. Latitude should be between -180 and 180'
+                            ),
+                        }
+                    ),
             })
             .or(
                 z.object({
@@ -40,7 +92,22 @@ export const organisationUnitSchema = identifiable
         programs: referenceCollection.optional().default([]),
         dataSets: referenceCollection.optional().default([]),
     })
+    .refine(
+        (orgUnit) => {
+            if (!orgUnit.id) {
+                return true
+            }
+            const isDescendantOfSelf = orgUnit.parent?.path.includes(orgUnit.id)
+            return !isDescendantOfSelf
+        },
+        {
+            message: i18n.t(
+                'Parent organisation unit cannot be itself or a descendant of itself.'
+            ),
+            path: ['parent'],
+        }
+    )
 
-export const initialValues = getDefaults(
-    organisationUnitSchema as z.AnyZodObject
-)
+export const initialValues = getDefaults(organisationUnitSchema)
+
+export const validate = createFormValidate(organisationUnitSchema)